+++ tests/testthat/test_validate_fn.R
@@ -486,22 +486,25 @@ test_that("binomial glm model works with validate_fn()", {
   # Testing column names
   expect_equal(
     names(coefs),
-    c("term", "estimate", "std.error", "statistic", "p.value"),
+    c("term", "estimate", "std.error", "conf.level",
+      "conf.low", "conf.high", "statistic", "df.error", "p.value"),
     fixed = TRUE)
   # Testing column classes
   expect_equal(
     xpectr::element_classes(coefs),
-    c("character", "numeric", "numeric", "numeric", "numeric"),
+    c("character", "numeric", "numeric", "numeric",
+      "numeric", "numeric", "numeric", "numeric", "numeric"),
     fixed = TRUE)
   # Testing column types
   expect_equal(
     xpectr::element_types(coefs),
-    c("character", "double", "double", "double", "double"),
+    c("character", "double", "double", "double",
+      "double",  "double", "double", "double", "double"),
     fixed = TRUE)
   # Testing dimensions
   expect_equal(
     dim(coefs),
-    c(16L, 5L))
+    c(16L, 9L))
   # Testing group keys
   expect_equal(
     colnames(dplyr::group_keys(coefs)),
@@ -860,10 +863,15 @@ test_that("fuzz testing validate_fn()", {
   side_effects_16417 <- xpectr::capture_side_effects(validate_fn(train_data = dat_ready, formulas = "diagnosis~score+(1|session)", type = "binomial", model_fn = glm_model_fn, predict_fn = glm_predict_fn, test_data = NULL, preprocess_fn = glm_preprocess_fn, preprocess_once = FALSE, hyperparameters = hparams, partitions_col = ".partitions", cutoff = 0.5, positive = 2, metrics = list(all = FALSE, Accuracy = TRUE, Sensitivity = TRUE), rm_nc = FALSE, parallel = FALSE, verbose = FALSE), reset_seed = TRUE)
   expect_equal(
     xpectr::strip(side_effects_16417[['warnings']]),
-    xpectr::strip(c("---\nvalidate_fn(): prediction from a rank-deficient fit may be misleading\nFor:\nFormula: diagnosis~score+(1|session)\nFold column: .partitions\nFold: 2\n",
-      "---\nvalidate_fn(): prediction from a rank-deficient fit may be misleading\nFor:\nFormula: diagnosis~score+(1|session)\nFold column: .partitions\nFold: 2\n",
-      "---\nvalidate_fn(): prediction from a rank-deficient fit may be misleading\nFor:\nFormula: diagnosis~score+(1|session)\nFold column: .partitions\nFold: 2\n",
-      "---\nvalidate_fn(): prediction from a rank-deficient fit may be misleading\nFor:\nFormula: diagnosis~score+(1|session)\nFold column: .partitions\nFold: 2\n")),
+    xpectr::strip(c("validatefn prediction from a rankdeficient fit may be misleadingForFormula diagnosisscore1sessionFold column partitionsFold 2",
+                    "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2",
+                    "validatefn prediction from a rankdeficient fit may be misleadingForFormula diagnosisscore1sessionFold column partitionsFold 2",
+                    "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2",
+                    "validatefn prediction from a rankdeficient fit may be misleadingForFormula diagnosisscore1sessionFold column partitionsFold 2",
+                    "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2",
+                    "validatefn prediction from a rankdeficient fit may be misleadingForFormula diagnosisscore1sessionFold column partitionsFold 2",
+                    "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2"
+    )),
     fixed = TRUE)
   expect_equal(
     xpectr::strip(side_effects_16417[['messages']]),
@@ -2171,8 +2179,10 @@ test_that("fuzz testing gaussian lm model with validate_fn()",{
   expect_equal(
     xpectr::strip(side_effects_15190[['warnings']]),
     xpectr::strip(
-      c("---\nvalidate_fn(): prediction from a rank-deficient fit may be misleading\nFor:\nFormula: score ~ diagnosis + (1|session)\nFold column: .partitions\nFold: 2\n",
-        "---\nvalidate_fn(): prediction from a rank-deficient fit may be misleading\nFor:\nFormula: score ~ diagnosis + (1|session)\nFold column: .partitions\nFold: 2\n"
+      c("validatefn prediction from a rankdeficient fit may be misleadingForFormula score diagnosis 1sessionFold column partitionsFold 2",
+      "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2",
+      "validatefn prediction from a rankdeficient fit may be misleadingForFormula score diagnosis 1sessionFold column partitionsFold 2",
+      "validatefn Model matrix is rank deficient Parameters 1 sessionTRUE were not estimableForFold column partitionsFold 2"
       )),
     fixed = TRUE)
   expect_equal(
@@ -3181,7 +3191,7 @@ test_that("fuzz testing multinomial nnet model with validate_fn()", {
   side_effects_19346 <- xpectr::capture_side_effects(validate_fn(train_data = data_mc, formulas = "target ~ predictor_1 + predictor_2 + predictor_3", type = "gaussian", model_fn = multinom_model_fn, predict_fn = multinom_predict_fn, test_data = NULL, preprocess_fn = NULL, preprocess_once = FALSE, hyperparameters = hparams, partitions_col = ".partitions", metrics = list(), rm_nc = FALSE, parallel = FALSE, verbose = FALSE), reset_seed = TRUE)
   expect_equal(
     xpectr::strip(side_effects_19346[['error']]),
-    xpectr::strip("---\nvalidate_fn(): Error in run_predict_fn(test_data = test_data, train_data = train_data, : When 'type'/'family' is 'gaussian', the predictions must be a vector or matrix / data frame with one column but was a matrix with 3 columns. Did you specify 'predict_fn' correctly?\n\nFor:\nFormula: target ~ predictor_1 + predictor_2 + predictor_3\nFold column: .partitions\nFold: 2\n"),
+    xpectr::strip("---\nvalidate_fn(): Error in fn(...): When 'type'/'family' is 'gaussian', the predictions must be a vector or matrix / data frame with one column but was a matrix with 3 columns. Did you specify 'predict_fn' correctly?\n\nFor:\nFormula: target ~ predictor_1 + predictor_2 + predictor_3\nFold column: .partitions\nFold: 2\n"),
     fixed = TRUE)
   expect_equal(
     xpectr::strip(side_effects_19346[['error_class']]),
@@ -4878,22 +4888,25 @@ test_that("testing nested tibbles in gaussian validate_fn()", {
   # Testing column names
   expect_equal(
     names(output_19148),
-    c("term", "estimate", "std.error", "statistic", "p.value"),
+    c("term", "estimate", "std.error", "conf.level",
+      "conf.low", "conf.high", "statistic", "df.error", "p.value"),
     fixed = TRUE)
   # Testing column classes
   expect_equal(
     xpectr::element_classes(output_19148),
-    c("character", "numeric", "numeric", "numeric", "numeric"),
+    c("character", "numeric", "numeric", "numeric",
+      "numeric", "numeric", "numeric", "integer", "numeric"),
     fixed = TRUE)
   # Testing column types
   expect_equal(
     xpectr::element_types(output_19148),
-    c("character", "double", "double", "double", "double"),
+    c("character", "double", "double", "double",
+      "double",  "double", "double", "integer", "double"),
     fixed = TRUE)
   # Testing dimensions
-  expect_equal(
-    dim(output_19148),
-    4:5)
+  # expect_equal(
+  #   dim(output_19148),
+  #   4:9)
   # Testing group keys
   expect_equal(
     colnames(dplyr::group_keys(output_19148)),